--- src/test/java/com/epam/tc/hw2/ex1/ConfProperties.java
@@ -1,35 +1,0 @@
-package com.epam.tc.hw2.ex1;
-
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.util.Properties;
-
-public class ConfProperties {
-    protected static FileInputStream fileInputStream;
-    protected static Properties PROPERTIES;
-
-    static {
-        try {
-            fileInputStream = new FileInputStream("src/test/resources/conf.properties");
-            PROPERTIES = new Properties();
-            PROPERTIES.load(fileInputStream);
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        } finally {
-            if (fileInputStream != null) {
-                try {
-                    fileInputStream.close();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-        }
-    }
-
-    public static String getProperty(String key) {
-        return PROPERTIES.getProperty(key);
-    }
-}
