--- src/test/java/com/epam/tc/hw2/ex1/GeneralWebDriver.java
@@ -1,8 +1,0 @@
-package com.epam.tc.hw2.ex1;
-
-import org.openqa.selenium.WebDriver;
-
-public class GeneralWebDriver {
-
-    public static WebDriver webDriver;
-}
