--- src/test/java/com/epam/tc/hw2/ex2/GeneralWebDriver.java
@@ -1,8 +1,0 @@
-package com.epam.tc.hw2.ex2;
-
-import org.openqa.selenium.WebDriver;
-
-public class GeneralWebDriver {
-
-    public static WebDriver webDriver;
-}
